CLEAN: refactored functions separated into modules
