+++ cleaning.py
@@ -1,29 +1,9 @@
-"""Script to load and clean life expectancy data
+"""Module to clean data in a DataFrame
 """
-import argparse
 import pandas as pd
 from pandas import DataFrame
 import numpy as np
 
-def main(region: str) -> None:
-    """Main function to load and clean life expectancy data
-
-    Args:
-        region: region to filter
-    """
-    df_raw = load_data()
-    df_clean = clean_data(region, df_raw)
-    save_data(df_clean)
-
-def load_data()-> DataFrame:
-    """Function to load life expectancy data
-    
-    Returns:
-        df_split: dataframe with split columns
-    """
-    df_raw = load_csv_file()
-    df_split = preprocess_data(df_raw)
-    return df_split
 
 def clean_data(region:str, data: DataFrame)-> DataFrame:
     """Function to clean life expectancy file, unpivot date column data 
@@ -48,40 +28,6 @@ def clean_data(region:str, data: DataFrame)-> DataFrame:
 
     return df_melt
 
-
-def save_data(df_clean: DataFrame)-> None:
-    """Function to save life expectancy data to csv file
-
-    Args:
-        df: dataframe to save
-    """
-    return save_file_to_csv(df_clean)
-
-
-def load_csv_file()-> DataFrame:
-    """Load a csv file into a DataFrame
-
-    Returns: a DataFrame containing the data from the csv file.
-    """
-    file_path = "data/eu_life_expectancy_raw.tsv"
-    return pd.read_csv(file_path, sep="\t")
-
-
-def preprocess_data(df_raw: DataFrame)-> DataFrame:
-    """Preprocesses raw data in a DataFrame by splitting a column and cleaning column names.
-
-    Args:
-        df_raw: The input DataFrame containing the raw data.
-
-    Returns:
-        df_split: A preprocessed DataFrame with split columns and cleaned column names.
-    """
-    df_raw[["unit", "sex","age", "region"]] = (
-        df_raw["unit,sex,age,geo\\time"].str.split(",", expand=True)
-    )
-    df_split = df_raw.drop(columns=["unit,sex,age,geo\\time"], inplace=False)
-    df_split.columns = df_split.columns.str.strip()
-    return df_split
 
 
 def identify_nans_convert_floats(df_split: DataFrame, value_cols: list)-> DataFrame:
@@ -138,23 +84,3 @@ def melt_dataframe(df_clean: DataFrame, id_cols: list, value_cols: list)-> DataF
         var_name="year"
     )
     return df_melt
-
-
-def save_file_to_csv(df_melt: DataFrame)-> None:
-    """ Save melted and clean dataframe to csv file
-
-    Args:
-        df_melt: dataframe to save
-    """
-    df_melt.to_csv(
-        "data/pt_life_expectancy.csv", 
-        sep=",",
-        index=False
-    )
-
-
-if __name__ == "__main__": #pragma: no cover
-    parser = argparse.ArgumentParser(description='Example argparse')
-    parser.add_argument("--region", "--r", type=str, default="PT", help='Region name')
-    args = parser.parse_args()
-    main(args.region)
