+++ tests/test_cleaning.py
@@ -1,13 +1,13 @@
 """Tests for the cleaning module"""
 import pandas as pd
 
-from ..cleaning import main
+from ..main import load_clean_save_data
 from . import OUTPUT_DIR
 
 
 def test_clean_data(pt_life_expectancy_expected):
     """Run the `clean_data` function and compare the output to the expected output"""
-    main(region="PT")
+    load_clean_save_data(region="PT")
     pt_life_expectancy_actual = pd.read_csv(
         OUTPUT_DIR / "pt_life_expectancy.csv"
     )
